Move internal packages to internal/
Move all packages that don't need to be used by clients to an internal/
directory so that their APIs can be more freely updated.

Add godoc comments to the remaining exported names under pkg/ that had
been missed before.
