+++ internal/algorithmregistry/algorithmregistry.go
