+++ internal/algorithmregistry/algorithmregistry_test.go
