+++ internal/tessera/safeint.go
