+++ tests/e2e_test.go
@@ -32,11 +32,11 @@ import (
 	v1 "github.com/sigstore/protobuf-specs/gen/pb-go/common/v1"
 	pbdsse "github.com/sigstore/protobuf-specs/gen/pb-go/dsse"
 	pbs "github.com/sigstore/protobuf-specs/gen/pb-go/rekor/v1"
+	"github.com/sigstore/rekor-tiles/internal/tessera"
 	"github.com/sigstore/rekor-tiles/pkg/client/read"
 	"github.com/sigstore/rekor-tiles/pkg/client/write"
 	pb "github.com/sigstore/rekor-tiles/pkg/generated/protobuf"
 	"github.com/sigstore/rekor-tiles/pkg/note"
-	"github.com/sigstore/rekor-tiles/pkg/tessera"
 	dsset "github.com/sigstore/rekor-tiles/pkg/types/dsse"
 	"github.com/sigstore/rekor-tiles/pkg/verify"
 	"github.com/sigstore/sigstore/pkg/signature"
